+++ app/Providers/EventServiceProvider.php
@@ -2,9 +2,6 @@
 
 namespace App\Providers;
 
-use App\Models\Category;
-use App\Models\Series;
-use App\Models\Volume;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;