+++ resources/views/livewire/volumes/create-volume.blade.php
@@ -17,7 +17,7 @@
                     </div>
                     <div class="row mt-1">
                         <div class="col-md-12">
-                            <label for="isbn" class="col-form-label required">{{ __('ISBN') }}</label>
+                            <label for="isbn" class="col-form-label">{{ __('ISBN') }}</label>
                             <div class="input-group">
                                 <input id="isbn" name="isbn" type="text" class="form-control @error('isbn') is-invalid @enderror" wire:model='isbn' autofocus>
                                 <button class="btn btn-outline-secondary" type="button" data-bs-toggle="modal" data-bs-target="#livestream_scanner"><i class="fa fa-barcode"></i></button>